changed phone_number to string

+++ db/migrate/20201223154506_remove_phone_number_from_forms.rb
@@ -1,5 +1,5 @@
 class RemovePhoneNumberFromForms < ActiveRecord::Migration[6.0]
   def change
-    remove_column :forms, :phone_number, :integer
+    remove_column :forms, :phone_number, :float
   end
 end
